Auto-generate Flow type definitions for graphql queries
Summary:
I got frustrated with writing out the types for graphql queries by hand, and decided to make the computer do it for us.
This depends on https://phabricator.khanacademy.org/D50334, which exposes an endpoint on webapp for updating our graphql schema.

This code (or something similar) could be used to fix https://khanacademy.atlassian.net/browse/MOB-694, which came out of the attempted/struggling outage earlier this year.

Test Plan: `yarn jest` should now generate flowtypes for all queries that are annotated with `# @autogen`. We can, in future, extend this to autogen for all queries, but I wanted to limit the change for now.

Reviewers: jeremy, bryan

Reviewed By: jeremy

Differential Revision: https://phabricator.khanacademy.org/D50367