Add support for GraphQL types that are marked as NON_NULL
Summary: Most of our GraphQL object's fields are marked as nullable, when they really aren't. So we can't trust them. But a few have been marked as NON_NULL. This adds support to our GraphQL response type generator for these.

Test Plan:
1. Add the `# @autogen` declaration to the `USER_COACHES_QUERY` in `ManagedCoaches/manage-coaches-api.js`
2. Run `yarn test queries`

There shouldn't be any errors and the generated file should correctly show the `nickname` field as a `string`.

Reviewers: jared

Reviewed By: jared

Subscribers: #mobile

Differential Revision: https://phabricator.khanacademy.org/D51700