Allow schema to be provided as a .graphql file, not just .json
Summary:
Webapp already has `genfiles/composed_schema.graphql`, which gqlgen is currently using do to type generation.
We could use the same! In fact I already had the logic for it in the test file 🎉

I did have to upgrade the graphql library to be able to parse some of the fancy stuff that webapp has though.

Test plan:
Checks pass, and try it out in webapp (PR in process).