Clean up index file generation & fix bug
Summary:
There was a bug where if there was one file that was the suffix of the other and both had queries ('ProfileWidgetsType.js' and 'getProfileWidgetsType.js' in the case I was looking at), then it would produce an invaid index file, with multiple entries for the one that was the suffix of the other.

I also took the opportunity to clean up some of the logic in this file, as it was a little hard to follow.

Test plan:
checks! Also run on webapp, see that index files don't contain duplicates.