Add support for generatedDir being an absolute path.
In this (currently) undocumented mode, the output goes into
```
    generatedDir/dirname/opname.js
```
rather than
```
    dirname/generatedDir/opname.js
```

This is useful if you want to keep all the output in a single tree
separate from the rest of your codebase.  I want to do this for some
validation tests, but I could see people wanting to do it if they have
read-only workspaces or something like that.

Unfortunately, `dirname` is always an absolute path currently, which
makes the output-directory structure depends on your own local
filesystem in ways that is not ideal.  I resolve this by making dirname relative to the current working directory, taking some care that the resulting output file can't escape the output generated-directroy.

Test plan:
I created a config file that looked like this:
```
{
	"schemaFilePath": ...,
	"options": {
		"generatedDirectory": "/tmp/gqlflow",
	}
}
```
and then ran
```
node ~/khan/graphql-flow/dist/cli/run.js /tmp/gqlflow/config.json
```
and saw output files like:
```
/tmp/gqlflow/services/static/javascript/promos-package/getPromoForUser.js
```
which is just what I was hoping to see.