Add ability to create directory config files (#40)
## Summary:
Users will be able to add files with the name `*graphql-flow.config.js` with a subset of the config fields (`options`, `excludes`) in order to have different behavior in that directory and its subdirectories.
Another field, `extends`, takes the path of another sub-config (or the root config) and will use the extended config as a base to add or replace option fields and add new excludes.
This is the first step in adding the ability to export robust flow enums, which we want to limit as to not disrupt other teams.

Issue: LP-11796

## Test plan:
Thoroughly unit tested. Add `graphql-flow.config.js` with different options to any subdirectory and re-run `graphql-flow`.

Author: nedredmond

Reviewers: jaredly

Required Reviewers:

Approved By: jaredly

Checks: ✅ Lint & Test (ubuntu-latest, 16.x)

Pull Request URL: https://github.com/Khan/graphql-flow/pull/40