Fix empty `variables` object to appease typescript-eslint
So, right now we're lazily using [flow-to-ts](https://github.com/Khan/flow-to-ts), which hasn't been updated in a while, and likely shouldn't be depended upon long-term.
The ~real ~better solution will be to just buckle down and produce typescript directly, instead of producing flow and then converting to typescript ... but that will be a battle for another day.