[🔥AUDIT🔥] Migrate repo to TypeScript (#60)
🖍 _This is an audit!_ 🖍

## Summary:
This is so that we can have proper types for TS nodes when updating the tool to generate TypeScript types directly.  The reason we want to do this is because we flow-to-ts breaks on IndexedAccessTypes and we can't update flow-to-ts because it forked babel.

Issue: None

## Test plan:
- yarn test
- yarn build
- yarn link
- cd ../webapp/services/static
- yarn link "@khanacademy/graphql-flow"
- cd ../..
- make gqlflow

Author: kevinbarabash

Auditors: jeresig, jaredly, somewhatabstract

Required Reviewers:

Approved By:

Checks: ✅ Lint & Test (ubuntu-latest, 16.x)

Pull Request URL: https://github.com/Khan/graphql-flow/pull/60